requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.10.0
elasticsearch==8.10.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
termcolor==2.3.0
urllib3==2.0.7
Werkzeug==3.0.1
